Remove a warning for bzero with GCC 12
GCC 12.2.0 on Arch reports:
.table.c: In function ???set_table_matrix0???:
.table.c:3536:5: error: ???__builtin_memset??? specified size between
18446744071562067969 and 18446744073709551615 exceeds maximum object
size 9223372036854775807 [-Werror=stringop-overflow=]
lto1: all warnings being treated as errors
lto-wrapper: fatal error: gcc returned 1 exit status
compilation terminated.
/usr/bin/ld: error: lto-wrapper failed
collect2: error: ld returned 1 exit status
child exited with value 1
make: *** [Makefile:132: w3m] Error 1

+++ table.c
@@ -3503,7 +3503,7 @@ correct_table_matrix4(struct table *t, int col, int cspan, char *flags,
 static void
 set_table_matrix0(struct table *t, int maxwidth)
 {
-    int size = t->maxcol + 1;
+    size_t size = t->maxcol + 1;
     int i, j, k, bcol, ecol;
     int width;
     double w0, w1, w, s, b;
